Instead of setting `display: block` or `display: none` on individual elements in the HTML, we refactor the code to use a CSS class `visible` instead such that we may control the style difference between a section being hidden or visible. The default, implicit state is hidden. In general it is a better approach to put as much styling information as possible in CSS files and keep the HTML mostly structural. The role of CSS is to specify the presentation. Separation of concerns is key.
Currently we use what is called a "CSS selector" to match our sections, the selector being `#imgtext > div`. The first part of the selector is `#imgtext` which means we select the element with id `imgtext`. We add `> div` to match all `<div>` elements that are direct children of `#imgtext`. To match the state where such elements also have a "visible" CSS class, we add `.visible`.

While currently we only toggle the `display` property as the class name changes, we are now in a position to make simple changes to the CSS file without touching either the HTML or JS file to augment the visual changes made during that state change.