Now we are in a good position to fix our bug.
First, let's think about our task here. We want our `switchToImage` function to hide
the previously-shown section and show the new section. This function gets the `id` of
the new section as its parameter. To be able to show the previous section, we should
keep track of the `id` of the previous section, so the first thing we'll do is to add
a global `selectedId` variable that will hold that state.

So the goal of our function can be split in 3 tasks:

1. hide the previous section (if it exists)
2. show the new section
3. keep track of the id of the newly-selected section

It's important when programming to make few assumptions on being provided valid
information. So we should also check that the `id` provided matches a valid section.

It's also very important when programming to do as little work as possible! So we
should also check that the newly-provided `id` is not the same as `selectedId`, since
in that case we have nothing to do. This could easily happen if the user clicks on the
same button multiple times.

We will still improve this code but we have now accomplished what we set out to do.