Sections are not visible by default and made visible by adding the `visible` CSS class. So far, we have implemented turning the visibility of sections on by simply adding `class = "visible"` and removing the `class` attribute altogether when a section is no longer visible.
However, we should strive to write code that does not make too many assumptions
on how the HTML is structured or codified. It's entirely possible that we'll want
to use CSS classes to denote more things for our sections, in which case it would
be possible for an element to have multiple CSS classes defined on it.

So instead of removing the `class` attribute or setting it to a single value, we
will use the element's `classList` property which offers methods on it to add and
remove a given class, with no impact on any pre-existing class.

It's also good practice to keep fixed values used in code, such as the name of our
`visible` CSS class in global constants. This will allow us to possibly use it in
multiple places, or change its name without having to look within functions, but
rather just look at the top of the file. This is a good code organization tip.